Fixed relocation errors
